Pull-Right class for columns to be able to float columns to the right instead of the left.
For example: a left sidebar should come after a main content area in the DOM, but using grids means you had to put it above. Now with this class, the sidebar is lower in the dom, but it will render left of the main content area if you use "pull-right" on the content area.